Move Matplotlib utilities to its own module
This allows to not have module constants in utils/misc.py that require
heavy imports.
