+++ spyder_kernels/console/start.py
@@ -18,8 +18,9 @@ import sys
 import site
 
 # Local imports
-from spyder_kernels.utils.misc import (
-    MPL_BACKENDS_FROM_SPYDER, INLINE_FIGURE_FORMATS, is_module_installed)
+from spyder_kernels.utils.misc import is_module_installed
+from spyder_kernels.utils.mpl import (
+    MPL_BACKENDS_FROM_SPYDER, INLINE_FIGURE_FORMATS)
 
 
 PY2 = sys.version[0] == '2'
