+++ spyder_kernels/utils/misc.py
@@ -11,25 +11,6 @@
 import re
 
 
-# Mapping of inline figure formats
-INLINE_FIGURE_FORMATS = {
-    '0': 'png',
-    '1': 'svg'
-}
-
-# Mapping of matlotlib backends options to Spyder
-MPL_BACKENDS_TO_SPYDER = {
-    'module://ipykernel.pylab.backend_inline': 0,
-    'Qt5Agg': 2,
-    'Qt4Agg': 3,
-    'MacOSX': 4,
-    'GTK3Agg': 5,
-    'GTKAgg': 6,
-    'WX': 7,
-    'TkAgg': 8
-}
-
-
 def is_module_installed(module_name):
     """
     Simpler version of spyder.utils.programs.is_module_installed.
@@ -45,33 +26,6 @@ def is_module_installed(module_name):
     except Exception:
         # Module is not installed
         return False
-
-
-def automatic_backend():
-    """Get Matplolib automatic backend option."""
-    if is_module_installed('PyQt5'):
-        auto_backend = 'qt5'
-    elif is_module_installed('PyQt4'):
-        auto_backend = 'qt4'
-    elif is_module_installed('_tkinter'):
-        auto_backend = 'tk'
-    else:
-        auto_backend = 'inline'
-    return auto_backend
-
-
-# Mapping of Spyder options to backends
-MPL_BACKENDS_FROM_SPYDER = {
-    '0': 'inline',
-    '1': automatic_backend(),
-    '2': 'qt5',
-    '3': 'qt4',
-    '4': 'osx',
-    '5': 'gtk3',
-    '6': 'gtk',
-    '7': 'wx',
-    '8': 'tk'
-}
 
 
 def fix_reference_name(name, blacklist=None):
